Moved TimeoutValidator to Polly.Timeout namespace
