+++ src/Polly.Shared/Timeout/TimeoutTResultSyntax.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using Polly.Timeout;
 using System.Threading.Tasks;
-using Polly.Utilities;
 
 namespace Polly
 {
